Move gmail and twitter clients
Also update references in utils Christmas tree imports
